Helheim temporary "fix"
Due to a series of Shiro Moments:tm:, Helheim cannot load all its custom objectives. Instead, it'll load in one of them.
Also, the reinforcements that come from the one extra objective will not be from Helheim, because of another Shiro Moment:tm:, and will instead come from Niflheim (which has the closest thing to undead foes).
I've messaged the developers that they have broken key aspects to the mod, hopefully they'll get back to me on this. I'll post any updates shortly.

The updated "temporarily fixed" version of Helheim is uploaded on the github.

I apologise for the inconvenience, my hands are sadly tied...